+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `--verify-proof` option to `import-object` command.
+
 ### Changed
 
 - Added support for Monero LWS v0.3.